Move common.evaluator_models to common.configs.evaluator
