+++ hasher-matcher-actioner/hmalib/common/tests/test_action_rule_evaluation.py
@@ -3,7 +3,7 @@
 import typing as t
 import unittest
 
-from hmalib.common.evaluator_models import (
+from hmalib.common.configs.evaluator import (
     ActionLabel,
     ActionRule,
 )
@@ -11,7 +11,6 @@ from hmalib.common.classification_models import (
     BankedContentIDClassificationLabel,
     BankIDClassificationLabel,
     ClassificationLabel,
-    Label,
 )
 from hmalib.common.message_models import BankedSignal, MatchMessage
 from hmalib.lambdas.actions.action_evaluator import get_actions_to_take
