+++ hasher-matcher-actioner/hmalib/lambdas/actions/action_evaluator.py
@@ -20,7 +20,7 @@ from hmalib.common.classification_models import (
     Label,
 )
 from hmalib.common.config import HMAConfig
-from hmalib.common.evaluator_models import (
+from hmalib.common.configs.evaluator import (
     ActionLabel,
     ActionRule,
 )
